microservices/chat-service/src/requirements.txt
Flask
Flask-PyMongo
gunicorn
python-dotenv
flask-restx